requirements.txt
Django==3.2.10
djangorestframework==3.12.4
requests==2.26.0